refactor: extract popover
