+++ apps/antalmanac/src/components/Calendar/calendar-event-popover.tsx
@@ -1,0 +1,58 @@
+import { Popover } from '@mui/material';
+import { useCallback, useEffect, useState } from 'react';
+import { shallow } from 'zustand/shallow';
+
+import CourseCalendarEvent from '$components/Calendar/CourseCalendarEvent';
+import AppStore from '$stores/AppStore';
+import { useSelectedEventStore } from '$stores/SelectedEventStore';
+
+export function CalendarEventPopover() {
+    const [anchorEl, selectedEvent, setSelectedEvent] = useSelectedEventStore(
+        (state) => [state.selectedEventAnchorEl, state.selectedEvent, state.setSelectedEvent],
+        shallow
+    );
+
+    const [scheduleNames, setScheduleNames] = useState(() => AppStore.getScheduleNames());
+
+    const handleClosePopover = useCallback(() => {
+        setSelectedEvent(null, null);
+    }, [setSelectedEvent]);
+
+    useEffect(() => {
+        const updateScheduleNames = () => {
+            setScheduleNames(AppStore.getScheduleNames());
+        };
+
+        AppStore.on('scheduleNamesChange', updateScheduleNames);
+
+        return () => {
+            AppStore.off('scheduleNamesChange', updateScheduleNames);
+        };
+    }, []);
+
+    if (!selectedEvent) {
+        return null;
+    }
+
+    return (
+        <Popover
+            anchorEl={anchorEl}
+            open={Boolean(anchorEl) && !!selectedEvent}
+            onClose={handleClosePopover}
+            anchorOrigin={{
+                vertical: 'bottom',
+                horizontal: 'left',
+            }}
+            transformOrigin={{
+                vertical: 'top',
+                horizontal: 'left',
+            }}
+        >
+            <CourseCalendarEvent
+                closePopover={handleClosePopover}
+                selectedEvent={selectedEvent}
+                scheduleNames={scheduleNames}
+            />
+        </Popover>
+    );
+}
